pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["Tamir Mandreev <tamirmandreev@mail.ru>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.5"
djangorestframework = "^3.16.1"
pytest = "^8.4.1"
pytest-django = "^4.11.1"
ipython = "^9.4.0"
pytest-cov = "^6.2.1"
python-dotenv = "^1.1.1"
drf-yasg = "^1.21.10"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'config.settings'


[tool.flake8]
max-line-length = 88
exclude = ['settings']


